Update values.schema.json
Improve the configurability of OpenWhisk in massive concurrent requests, because when the upper limit of the configuration is tens of millions, the yaml file of the pod generated after helm installation will be wrongly expressed as scientific notation, so we need to avoid this unexpected failure. The above .
so, I suggest change string type.